10-2 文档页面和文档页面管理，只能查询当前电子书下的所有文档

+++ kb/src/main/java/com/song/kb/domain/DocExample.java
@@ -484,7 +484,7 @@ public class DocExample {
             return (Criteria) this;
         }
 
-        public Criteria andCoteCountEqualTo(Integer value) {
+        public Criteria andCoteCountEqualTo(Long value) {
             addCriterion("cote_count =", value, "coteCount");
             return (Criteria) this;
         }
